refactor: . move new  functions in QueryRendererHelper.ts to Postponer.ts
