+++ src/lib/QueryRendererHelper.ts
@@ -1,9 +1,6 @@
-import type { Moment, unitOfTime } from 'moment';
 import type { GlobalFilter } from '../Config/GlobalFilter';
 import type { GlobalQuery } from '../Config/GlobalQuery';
 import { Query } from '../Query/Query';
-import { Task } from '../Task';
-import { TasksDate } from '../Scripting/TasksDate';
 
 /**
  * @summary
@@ -70,40 +67,4 @@ export function getQueryForQueryRenderer(source: string, globalQuery: GlobalQuer
     }
 
     return globalQuery.query(path).append(tasksBlockQuery);
-}
-
-export type HappensDate = keyof Pick<Task, 'startDate' | 'scheduledDate' | 'dueDate'>;
-
-/**
- * Gets a {@link HappensDate} field from a {@link Task} with the following priority: due > scheduled > start.
- * If the task has no happens field {@link HappensDate}, null is returned.
- *
- * @param task
- */
-export function getDateFieldToPostpone(task: Task): HappensDate | null {
-    if (task.dueDate) {
-        return 'dueDate';
-    }
-
-    if (task.scheduledDate) {
-        return 'scheduledDate';
-    }
-
-    if (task.startDate) {
-        return 'startDate';
-    }
-
-    return null;
-}
-
-export function createPostponedTask(
-    task: Task,
-    dateTypeToUpdate: HappensDate,
-    timeUnit: unitOfTime.DurationConstructor,
-    amount: number,
-) {
-    const dateToUpdate = task[dateTypeToUpdate] as Moment;
-    const postponedDate = new TasksDate(dateToUpdate).postpone(timeUnit, amount);
-    const newTasks = new Task({ ...task, [dateTypeToUpdate]: postponedDate });
-    return { postponedDate, newTasks };
 }
